requirements.txt
#Python 3.10.11
fastapi
python-jose
passlib
pymongo
python-multipart
deta